refactor(core): extract TranscribeRequest and overrides
Move TranscribeSource / TranscribeRequest into types/request.rs
and hotword / override types into types/overrides.rs. Public
gigastt_core::inference::* paths are unchanged.
